+++ torii-services/Cargo.toml
@@ -1,18 +1,19 @@
 [package]
-name = "torii-core"
-description = "Core functionality for the torii authentication ecosystem"
+name = "torii-services"
+description = "Service implementations for the torii authentication ecosystem"
 version = "0.5.2"
 edition.workspace = true
 repository.workspace = true
 license.workspace = true
 
 [dependencies]
+torii-core = { path = "../torii-core", version = "0.5.2" }
 async-trait.workspace = true
 base64.workspace = true
 chrono.workspace = true
 dashmap.workspace = true
 downcast-rs = "2.0.1"
-jsonwebtoken = { workspace = true, optional = true }
+jsonwebtoken.workspace = true
 password-auth = "1.0.0"
 sha2 = "0.10"
 subtle = "2.5"
@@ -24,14 +25,12 @@ serde_json.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 uuid.workspace = true
+tokio.workspace = true
 b58 = "0.1.2"
 
 # Optional mailer integration
 torii-mailer = { path = "../torii-mailer", version = "0.5.2", optional = true }
 secrecy = "0.10.3"
-
-[target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
-getrandom = { version = "0.3", features = ["wasm_js"] }
 
 [dev-dependencies]
 tokio.workspace = true
@@ -39,5 +38,3 @@ tokio.workspace = true
 [features]
 default = []
 mailer = ["dep:torii-mailer"]
-jwt = ["dep:jsonwebtoken"]
-wasm = ["uuid/js"]
